feat: Can load up the server now and view app on localhost
