--- style.css
@@ -1,33 +1,0 @@
-h1,h5,p,button { 
-    font-family: monaco,Consolas,Lucida Console,monospace;
-}
-
-.btn:focus {
-    outline: none;
-    box-shadow: none;
-  }
-
-
-#canvas {
-    margin: 20px;
-}
-
-.color-picker{
-  font-size:24px;
-}
-
-.colorPicker {
-  position: relative;
-  background: rgba(0, 0, 0, 0);
-  text-align: center;
-}
-
-.colorPickerInput {
-  position: absolute;
-  top: -5px;
-  left: -5px;
-  height: 110%;
-  width: 110%;
-  opacity: 0;
-  cursor: pointer;
-}
