--- view.html
@@ -1,158 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <title>Etch-A-Sketch</title>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1 user-scalable=no">
-  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
-  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
-  <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.11.8/dist/umd/popper.min.js" integrity="sha384-I7E8VVD/ismYTF4hNIPjVp/Zjvgyol6VFvRkX/vR+Vc4jQkC+hVqc2pM8ODewa9r" crossorigin="anonymous"></script>
-  <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.5.0/font/bootstrap-icons.css">
-  <link rel="stylesheet" href="style.css">
-
-</head>
-<body>
-
-<div class="container-fluid text-center">
-  <!-- This is the main title text -->
-  <div class="row">
-        <h1>Etch-A-Sketch</h1>
-  </div>
-  <!-- This is the row of buttons to control the colors and other options -->
-  <div class="row">
-      <div class="col justify-content-center">
-      <button type="button" id="blackButton" class="btn btn-outline-dark mx-2 my-1">Black</button>
-      <button type="button" id="rainbowButton" class="btn btn-outline-dark mx-2 my-1">Rainbow</button>
-      <!-- Button for color picker. We set the opacity for the actual color picker to zero so it appears as a button -->
-      <button type="button" class="btn btn-outline-dark mx-2 my-1">
-        <div class="colorPicker">Pick A Color
-          <input class="colorPickerInput" type="color" id="color"/>
-        </div>
-      </button>
-      <button type="button" id="eraserButton" class="btn btn-outline-dark mx-2 my-1">Eraser</button>
-      <button type="button" id="clearButton" class="btn btn-outline-dark mx-2 my-1">Clear</button>
-      <button class="btn btn-outline-dark mx-2 my-1 dropdown-toggle" type="button" id="dropdownMenuButton1" data-bs-toggle="dropdown" aria-expanded="false">
-          More
-      </button>
-        <ul class="dropdown-menu" aria-labelledby="dropdownMenuButton1">
-          <!-- Modal for size -->
-          <li><a aria-pressed="true" data-bs-toggle="modal" data-bs-target="#sizeModal" class="dropdown-item" href="#">Size</a></li>
-          <li><a class="dropdown-item" id="borderlessButton" href="#">Borderless</a></li>
-          <li><a class="dropdown-item" href="#">Save</a></li>
-          <li><a class="dropdown-item" href="#">Help</a></li>
-          <!-- Change text if already logged in -->
-          <li><a data-bs-toggle="modal" data-bs-target="#loginModal" class="dropdown-item" href="#">Login</a></li>
-        </ul>
-      </div>
-  </div>
-
-  <!-- This is the main drawing grid itself. -->
-  <div class="row">
-    <div class="col">
-      <canvas id="canvas" width="620px" height="620px"></canvas>
-    </div>
-  </div>
-  
-</div>
-
-<div>
-    <footer class="container-fluid text-center text-white mt-4" style="background-color: #000000">
-        <p>Made by Kunal Varkekar. 2023.</p>
-    </footer>
-</div>
-
-<!-- Size Modal -->
-<div class="modal fade" id="sizeModal" tabindex="-1" aria-labelledby="sizeModalLabel" aria-hidden="true">
-  <div class="modal-dialog modal-dialog-centered">
-    <div class="modal-content">
-      <div class="modal-header">
-        <h5 class="modal-title" id="sizeModalLabel">Enter the number of boxes:</h5>
-        <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-      </div>
-      <div class="modal-body">
-        <form id="sizeForm">
-          <div class="mb-3">
-            <label for="number_boxes" class="col-form-label">Warning: This will erase all your current art.</label>
-            <input type="number" value="15" class="form-control" id="number_boxes">
-          </div>
-          <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Cancel</button>
-        <button type="button" id="sizeButton" data-bs-dismiss="modal" class="btn btn-primary">Submit</button>
-        </form>
-      </div>
-    </div>
-  </div>
-</div>
-
-<!-- Login Modal -->
-<div class="modal fade" id="loginModal" tabindex="-1" aria-labelledby="loginModalLabel" aria-hidden="true">
-  <div class="modal-dialog modal-dialog-centered">
-    <div class="modal-content">
-      <div class="modal-header">
-        <h5 class="modal-title" id="loginModalLabel">Login to your account</h5>
-        <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-      </div>
-      <div class="modal-body">
-        <form class="px-2 py-2" id="loginForm">
-          <div class="mb-3">
-            <label for="username" class="col-form-label">Username</label>
-            <input type="text" placeholder="Username" class="form-control" id="username">
-          </div>
-            <div class="mb-3">
-            <label for="password" class="col-form-label">Password</label>
-            <input type="text" placeholder="Password" class="form-control" id="password">
-          </div>
-            <div class="mb-3">
-            <input type="checkbox" class="form-check-input" id="dropdownCheck">
-        <label class="form-check-label" for="dropdownCheck">
-          Remember me
-        </label>
-          </div>
-          
-          <button type="button" id="loginButton" data-bs-dismiss="modal" class="btn btn-primary">Login</button>
-        </form>
-        
-        <div class="list-group list-group-flush">
-          <a href="#" data-bs-toggle="modal" data-bs-target="#signupModal" data-dismiss="modal" class="list-group-item list-group-item-action">New around here? Sign up!</a>
-          <a href="#" class="list-group-item list-group-item-action" >Forgot Password</a>
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
-
-<!-- Signup Modal -->
-<div class="modal fade" id="signupModal" tabindex="-1" aria-labelledby="signupModalLabel" aria-hidden="true">
-  <div class="modal-dialog modal-dialog-centered">
-    <div class="modal-content">
-      <div class="modal-header">
-        <h5 class="modal-title" id="signupModalLabel">Signup for an account</h5>
-        <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-      </div>
-      <div class="modal-body">
-        <form class="px-2 py-2" id="signupForm">
-          <div class="mb-3">
-            <label for="email" class="col-form-label">Email</label>
-            <input type="text" placeholder="Email" class="form-control" id="email">
-          </div>
-          <div class="mb-3">
-            <label for="username" class="col-form-label">Username</label>
-            <input type="text" placeholder="Username" class="form-control" id="username">
-          </div>
-          <div class="mb-3">
-            <label for="password" class="col-form-label">Password</label>
-            <input type="text" placeholder="Password" class="form-control" id="password">
-          </div>
-          <button type="button" id="signupButton" data-bs-dismiss="modal" class="btn btn-primary">Sign Up</button>
-        </form>
-        <div class="list-group list-group-flush">
-          <a href="#" data-bs-toggle="modal" data-bs-target="#loginModal" data-dismiss="modal" class="list-group-item list-group-item-action">Already signed up? Log in!</a>
-          <a href="#" class="list-group-item list-group-item-action" >Forgot Password</a>
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
-
-</body>
-<script src="grid.js"></script>
-</html>
